requirements.txt
beautifulsoup4
crawl4ai
fastapi
Markdown
pydantic
uvicorn
python-dotenv
loguru
faiss-cpu
langchain
langchain-text-splitters
langchain-groq
langchain-core
langchain-community
langchain-huggingface
langchain-groq
requests
weasyprint
sentence-transformers
streamlit
markdown-pdf
lxml 
python-pptx 
md2pdf